% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{clean geocode}
\usage{
cleanData(data = NULL, path, header = F, name = "sanhuang",
  direction = "nsz", sam = T, npoint = 4, from = 3, to = 5)
}
\arguments{
\item{data}{data from the other source(gaodemap)}

\item{path}{data path to read}

\item{header}{logical,read.csv(header=F)}

\item{name}{character,data is road name}

\item{direction}{character,data's direction}

\item{sam}{logical while true,will sample the color of gecodata}

\item{npoint}{numeric sample num}

\item{to}{5 or 6. See more in details.}
}
\description{
clean geocode from gaodemap to a list. this list has two dataframe ,one is
point data,another is line data.
}
\details{
Go to http://developer.baidu.com/map/index.php?title=webapi/guide/changeposition to see what the intergers mean.
}

