% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoconv.R
\name{geoconv}
\alias{geoconv}
\title{Convert geocode}
\usage{
geoconv(geocode, from = 3, to = 5)
}
\arguments{
\item{from}{takes intergers from 1 to 8. See more in details.}

\item{to}{5 or 6. See more in details.}

\item{geocde}{geocode from the other source,limit 100 geo point once}
}
\description{
Take in geocode from the other source to baidu's geocode
}
\details{
Go to http://developer.baidu.com/map/index.php?title=webapi/guide/changeposition to see what the intergers mean.
}
\examples{

vec=geoconv(geocode=c(119.348557,25.996298),from=3,to=5)

mtx=matrix(c(119.348557,25.996298,119.350449,25.997036),byrow=T,ncol=2)
geo_mtx=geoconv(mtx)

df=as.data.frame(mtx)
geo_df=geoconv(df)
}

