% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoconvA.R
\name{geoconvA}
\alias{geoconvA}
\title{translate locations by amap API}
\usage{
geoconvA(locations = "116.481499,39.990475|116.481499,39.990375",
  coordsys = "baidu", mapkey, mtx = T)
}
\arguments{
\item{locations}{a vector a or matrix contains longtitude and latitude of the center of the map, or a character refers to the address}

\item{coordsys}{width of the map}

\item{mapkey}{height of the map}

\item{mtx}{multiplicative factor for the number of pixels returned. possible values are 2 or anything else.}
}
\description{
Take in coordiantes and return the location changed by amap API
}
\examples{

\dontrun{
amap_key='71b002162d43fea08afd5b5e31b030af'
geoconvA(mapkey = amap_key)
geoconvA(mapkey = amap_key,mtx=F)
}
}

