% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoordinateA.R
\name{getCoordinateA}
\alias{getCoordinateA}
\title{get Coordinate from amap}
\usage{
getCoordinateA(address = NULL, city = "¸£Öİ", key = NULL)
}
\arguments{
\item{address}{character:defalut NULL ,should be rectangle/circle/NULL}

\item{city}{character:location should be writtend like}

\item{key}{should be 'json' or 'xml', the type of the result}
}
\value{
A vector contains description,expedite,congested,blocked,unkown
}
\description{
get traffic status from amap,there are c(description,expedite,congested,blocked,unkown)
}
\examples{
\dontrun{
## rectangle
amap_key='71b002162d43fea08afd5b5e31b030af'
getCoordinateA(address='ç‘—å—æ£¬ç»”<U+393C><U+3E39>',city='fuzhou',key=amap_key)
MapBarBus$location=NA
for(i in 1:nrow(MapBarBus)){
  MapBarBus$location[i]=getCoordinateA(address=MapBarBus$station[i],city='ç»‚å¿“çª<U+393C><U+3E65>',key=amap_key)
}

}
}

