% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMapbarBus.R
\name{getMapBarBus}
\alias{getMapBarBus}
\title{get MapBar Bus from 'http://bus.mapbar.com/'}
\usage{
getMapBarBus(url = "http://bus.mapbar.com/fuzhou1/xianlu/", Ch = NULL,
  Py = c("ditie", "puwei", "guang", "xiang"))
}
\arguments{
\item{url}{address}

\item{Ch}{cause tmcn::toPinyin's bug,some chinese word need to be changed to pinyin}

\item{Py}{pinyin}
}
\value{
df  bus station
}
\description{
get MapBar Bus from 'http://bus.mapbar.com/'
}
\examples{
df=getMapBarBus(url='http://bus.mapbar.com/fuzhou1/xianlu/',
 Ch=c('åœ°é“','åŸ”å°¾','å¹‘ù<U+3E66>','å·‘ù<U+3E37>'),
 Py=c('ditie','puwei','guang','xiang'))


 df=getMapBarBus()
}

