% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlace.R
\name{getPlace}
\alias{getPlace}
\title{Get place from query
Take in query and city, return the informations}
\usage{
getPlace(place = NULL, city = "±±¾©", page_size = 20, pages = Inf,
  scope = 1, verbose = TRUE, bounds = NULL)
}
\arguments{
\item{place}{the place you want to search}

\item{city}{define the city}
}
\value{
a data frame contains name, longtitude, latitude and address, as well as teleplhone number, if exist.
}
\description{
Get place from query
Take in query and city, return the informations
}
\examples{
\dontrun{
## colleges in beijing
bj_college = getPlace('æ¾¶Ñƒî„Ÿ', 'é–æ¤¾åõ<U+3E62>')
## Mcdonald's in shanghai
sh_mcdonald = getPlace('æ¥¹ï¹€ç¶‹é”æˆ¼ã¸³', 'æ¶“å©ƒæõ<U+3E33>')
}
}

