% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoadCoor.R
\name{getRoadCoor}
\alias{getRoadCoor}
\title{get road ployline coordinate by road id from amap}
\usage{
getRoadCoor(id = "B0FFGQOZO5", tobaidu = T, name = "hhh")
}
\arguments{
\item{id}{road'id,can get by function getSearchPlus}

\item{tobaidu}{whether convert amap to baidu coodinate,default to be T}
}
\value{
a list within linedata and geodata
}
\description{
get road ployline from amap,default will convert to baidu coordinate,and will
convert to linedata and pointdata to show within REmap::remapb()
}
\examples{
\dontrun{
rectangel=matrix(c(120.208958,27.327244,120.232389,27.30612),byrow = T,nrow = 2) #test
df=getSearchPlus(mtx = rectangel,word = 'é–¬æ’¹çŸ‘ù<U+3E65>',types = '19',Tobaidu = F)
str(df)
x=getroadcoor()
 }
}

