% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearch.R
\name{getSearch}
\alias{getSearch}
\title{search something from amap}
\usage{
getSearch(shape = "polygon", location = "116.456299,39.960767",
  keywords = NULL, offset = 25, polygon = NULL, radius = 10000,
  city = NULL, key = amap_key, extensions = "base",
  types = "地名地址信息")
}
\arguments{
\item{shape}{character:defalut NULL ,should be rectangle/circle/NULL}

\item{location}{character:location should be writtend like}

\item{keywords}{should be 'json' or 'xml', the type of the result}

\item{offset}{code of city.for examples fuzhou is 0591}

\item{polygon}{polygon of location}

\item{radius}{logical value, return the coordinates or the original results}

\item{city}{character}

\item{key}{amap key}

\item{extensions}{xml or json}
}
\value{
A vector contains description,expedite,congested,blocked,unkown
}
\description{
search something from amap,you can see http://lbs.amap.com/api/webservice/guide/api/trafficstatus/#inputtips
}
\examples{
\dontrun{
rectangel=matrix(c(118.784364,26.478515,119.892801,25.436929),byrow = T,nrow = 2) #big
rectangel=matrix(c(119.117295,26.201212,119.649092,25.840692),byrow = T,nrow = 2) #saml
df_grid=setcoordMTX(rectangel=rectangel,grid=2000,sep=';')
}
}

