% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearch.R
\name{getSearchPlus}
\alias{getSearchPlus}
\title{search something from amap}
\usage{
getSearchPlus(mtx, grid = 5000, word = "bicycle", Tobaidu = T,
  types = "地名地址信息")
}
\arguments{
\item{mtx}{character:defalut NULL ,should be rectangle/circle/NULL}

\item{grid}{character:location should be writtend like}

\item{word}{should be 'json' or 'xml', the type of the result}

\item{Tobaidu}{code of city.for examples fuzhou is 0591}
}
\description{
search something from amap,you can see http://lbs.amap.com/api/webservice/guide/api/trafficstatus/#inputtips
}

