% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrafficStatus.R
\name{getTrafficStatus}
\alias{getTrafficStatus}
\title{Get Traffic Status from amap}
\usage{
getTrafficStatus(type = "rectangle", location = NULL, name = NULL,
  radius = NULL, city = NULL, citycode = "350100", key = amap_key,
  output = "XML")
}
\arguments{
\item{type}{character:defalut NULL ,should be rectangle/circle/NULL}

\item{location}{character:location should be writtend like}

\item{name}{should be 'json' or 'xml', the type of the result}

\item{radius}{logical value, return the coordinates or the original results}

\item{city}{character}

\item{citycode}{code of city.for examples fuzhou is 0591}

\item{key}{amap key}

\item{output}{xml or json}
}
\value{
A vector contains description,expedite,congested,blocked,unkown
}
\description{
get traffic status from amap,there are c(description,expedite,congested,blocked,unkown)
}
\examples{
\dontrun{
## rectangle
location='119.278861,26.106208;119.336454,26.078304'
# the rectangle can't be too big,diagonal must be shorted than 10km
a=getTrafficStatus(location = location,key = amap_key,type='rectangle')

## circle
location='119.278861,26.106208'
b=getTrafficStatus(type='circle',location = location,key = amap_key)

## road
location='119.278861,26.106208'
c=getTrafficStatus(type='name',key = amap_key,name='鍙ょ敯璺愼㹦')
}
}

