% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrafficStatus.R
\name{getTrafficStatusPlus}
\alias{getTrafficStatusPlus}
\title{Get Traffic Status from amap from grid cell}
\usage{
getTrafficStatusPlus(location = NULL, key = amap_key, grid = 5000)
}
\arguments{
\item{location}{character:location should be writtend like}

\item{key}{amap key}

\item{citycode}{code of city.for examples fuzhou is 0591}

\item{gird}{sides of rectangle,default 5000m}
}
\value{
A vector contains description,expedite,congested,blocked,unkown
}
\description{
get traffic status from amap,there are c(description,expedite,congested,blocked,unkown)
}
\examples{
\dontrun{
location='119.229335,26.122166;119.59875,25.87036'
df=getTrafficStatusPlus(location = location)
}
}

