% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSearch.R
\name{geturl0}
\alias{geturl0}
\title{get search API url}
\usage{
geturl0(page = 1, shape = "around", location = "116.456299,39.960767",
  keywords = NULL, offset = 25, polygon = NULL, radius = 2000,
  city = NULL, mapkey = "71b002162d43fea08afd5b5e31b030af",
  types = "µØÃûµØÖ·ĞÅÏ¢", extensions = "base")
}
\arguments{
\item{page}{num}

\item{shape}{character:defalut NULL ,should be rectangle/circle/NULL}

\item{location}{character:location should be writtend like}

\item{keywords}{should be 'json' or 'xml', the type of the result}

\item{offset}{code of city.for examples fuzhou is 0591}

\item{polygon}{polygon of location}

\item{radius}{logical value, return the coordinates or the original results}

\item{city}{character}

\item{mapkey}{amap key}

\item{extensions}{xml or json}
}
\value{
A vector contains description,expedite,congested,blocked,unkown
}
\description{
search something from amap,you can see http://lbs.amap.com/api/webservice/guide/api/trafficstatus/#inputtips
}
\examples{
\dontrun{
rectangel=matrix(c(119.117295,26.201212,119.649092,25.840692),byrow = T,nrow = 2)
df_grid=setcoordMTX(rectangel=rectangel,grid=5000,sep=';')
amap_key='71b002162d43fea08afd5b5e31b030af'
url=geturl0(shape='polygon',polygon=df_grid[1,1],mapkey=amap_key,location=NULL,keywords='éîƒ¿æ°õ<U+3E36>??<U+393C><U+3E39>')
url=geturl0(shape='polygon',polygon=df_grid[1,1],mapkey=amap_key,location=NULL,keywords='éîƒ¿æ°¦ç»”æª¤éîƒ¿æ°¦æï¸¾çõ<U+3E66>')
url=geturl0(shape='polygon',polygon=df_grid[1,1],mapkey=amap_key,location=NULL,keywords='é–¬æ’¹çŸ‘ù<U+3E65>',types='é¦æ¿æ‚•é¦æ¿æ½ƒæ·‡â„ƒä¼<U+383C><U+3E35>')
 }
}

