% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRouteA.R
\name{geturlRouteA}
\alias{geturlRouteA}
\title{Get Traffic Status from amap}
\usage{
geturlRouteA(orgin = "116.481499,39.990475",
  destination = "116.465063,39.999538", city = "beijing", strategy = 2,
  mapkey = "71b002162d43fea08afd5b5e31b030af")
}
\arguments{
\item{city}{character}

\item{type}{character:defalut NULL ,should be rectangle/circle/NULL}

\item{location}{character:location should be writtend like}

\item{name}{should be 'json' or 'xml', the type of the result}

\item{radius}{logical value, return the coordinates or the original results}

\item{citycode}{code of city.for examples fuzhou is 0591}

\item{key}{amap key}

\item{output}{xml or json}
}
\value{
A vector contains description,expedite,congested,blocked,unkown
}
\description{
get traffic status from amap,there are c(description,expedite,congested,blocked,unkown)
}
\examples{
\dontrun{
## rectangle
location='119.278861,26.106208;119.336454,26.078304'
a=getTrafficStatus(location = location,key = amap_key,type='rectangle')

## circle
location='119.278861,26.106208'
b=getTrafficStatus(type='circle',location = location,key = amap_key)

## road
location='119.278861,26.106208'
c=getTrafficStatus(type='name',key = amap_key,name='鍙ょ敯璺㤼㸰?<U+3E66>')
}
}

