% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poi2Loi.R
\name{loc2m}
\alias{loc2m}
\title{transform Lontitude and latitude to meter}
\usage{
loc2m(mtx)
}
\arguments{
\item{mtx}{data to transform}
}
\description{
transform Lontitude and latitude to meter
}
\examples{
\dontrun{
xy=matrix(c(119.328934,26.101757,119.328957,26.101746),ncol=2,byrow=T)
loc2m(xy)
}

}

