% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline2mtx.R
\name{mtx2geoData}
\alias{mtx2geoData}
\title{mtx2geoData}
\usage{
mtx2geoData(mtx, name = "point")
}
\arguments{
\item{mtx}{matrix}

\item{name}{character}
}
\value{
data.frame
}
\description{
matirx data to geoData to use remapB
}
\examples{
\dontrun{
geoData=mtx2geoData(mtx)
}


}

