% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline2mtx.R
\name{mtx2lineData}
\alias{mtx2lineData}
\title{mtx2lineData}
\usage{
mtx2lineData(mtx, name = "point", color = NULL)
}
\arguments{
\item{mtx}{matrix}

\item{name}{character}

\item{color}{character,used to set line theme}
}
\value{
data.frame
}
\description{
matirx data to lineData to use remapB
}
\examples{
\dontrun{
lineData=mtx2lineData(mtx,color ='blue')
}


}

