% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poi2Loi.R
\name{Poi2Loi}
\alias{Poi2Loi}
\title{change POI data to LOI(line of interest)}
\usage{
Poi2Loi(data, strsplitcol = "address", splitcha = "[;/.]")
}
\arguments{
\item{data}{data from the other source(gaodemap)}

\item{strsplitcol}{data path to read}

\item{splitcha}{logical,read.csv(header=F)}

\item{to}{5 or 6. See more in details.}
}
\description{
change bus station poi to bus route LOI
}
\examples{

\dontrun{
data(bus)
df=Poi2Loi(bus,strsplitcol='address',splitcha='[;/.]')
}
}

