% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyline2mtx.R
\name{polyline2mtx}
\alias{polyline2mtx}
\title{polyline2mtx}
\usage{
polyline2mtx(polyline, split = ";")
}
\arguments{
\item{polyline}{polyline location}
}
\value{
matrix
}
\description{
polyline location to matrix type
}
\examples{
\dontrun{
mtx=polyline2mtx(routeout)
}


}

