% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poi2Loi.R
\name{Roadgrid}
\alias{Roadgrid}
\title{transform Lontitude and latitude to meter}
\usage{
Roadgrid(mtx, grid = 50)
}
\arguments{
\item{mtx}{colnames(mtx)=c('lng','lat')}
}
\description{
transform Lontitude and latitude to meter
}
\examples{
\dontrun{
mtx=matrix(c(119.328934,26.101757,119.328957,26.101746),ncol=2,byrow=T)
mtx0=Roadgrid(mtx)
}

}

