% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setcoordMTX.R
\name{setcoordMTX}
\alias{setcoordMTX}
\title{cut map rectangel to small cell}
\usage{
setcoordMTX(rectangel, grid = 5000, sep = ";")
}
\arguments{
\item{rectangel}{matrix:}

\item{grid}{numeic: samll cell length}

\item{sep}{see detail}
}
\value{
a data.frame
}
\description{
for example rectangel=matrix(c(119.232566,26.131948,119.47288,25.941584),byrow = T,nrow = 2),
but amap and baidumap can only search POI in a small cell ,so we need to cut big
one to many samll cell
}
\examples{
\dontrun{
## rectangle

rectangel=matrix(c(119.281777,26.08224,119.324092,26.059111),byrow = T,nrow = 2)
df_grid=setcoordMTX(rectangel=rectangel,grid=5000,sep=';')
}
}

